Fix some type cast

+++ SelectFileDlg.cpp
@@ -80,7 +80,7 @@ void CSelectFileDlg::OnBtnExploreXSDFile() {
             std::ifstream t((LPCTSTR)ret);
 
             t.seekg(0, std::ios::end);
-            str.reserve(t.tellg());
+            str.reserve((size_t) t.tellg());
             t.seekg(0, std::ios::beg);
 
             str.assign((std::istreambuf_iterator<char>(t)), std::istreambuf_iterator<char>());
